requirements.txt
BeautifulSoup
requests==2.7.0
Pillow==2.9.0
